LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/mrow/inferred-mrow-stretchy-expected.txt


PASS operator stretching inside Mrow
PASS operator stretching inside Sqrt
PASS operator stretching inside Style
PASS operator stretching inside Error
PASS operator stretching inside Phantom
PASS operator stretching inside Math
PASS operator stretching inside Menclose
PASS operator stretching inside Mpadded
FAIL operator stretching inside Unknown assert_greater_than_equal: expected a number greater than or equal to 100 but got 0
↨

↨

↨

↨


↨

↨

↨
